# perf: precompile contracts for faster test runs