--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,9 @@
 # Foundry build and cache directories
 out/
+out-optimized/
 cache/
 node_modules/
+
+# Coverage
+report/
+lcov.info