## Motivation

IR compilation of the entire project takes a long time. When excluding the test suite however, the time cuts down dramatically. It would be nice to compile just the source contracts via IR and the test suite without. However, there's currently no way (in Foundry) to specify different profiles or different compilation configuration for specific files in a build.

## Solution

The approach taken here is inspired by https://github.com/ProjectOpenSea/seaport. I've created a profile `optimized-build` for the purpose of precompiling just the source contracts via IR into the folder `out-optimized`. When tests are run under the `optimized-test` or `optimized-test-deep` profiles, the project is built again without IR, but the relevant source contracts are deployed using the optimized bytecode from `out-optimized` so the tests can test against the bytecode that is actually to be deployed.

I've updated the CI workflow to remove the lite test in favor of using the optimized test. It should be fast enough now that we don't need the early signal provided by the lite profile.

|Before (3m 59s)|After (58s)|
|---|---|
|![Screenshot 2023-11-29 at 1 27 29 AM](https://github.com/erc6900/reference-implementation/assets/8624213/b46b2052-d9e9-4827-aeca-da3f18dfab15)|![Screenshot 2023-11-29 at 1 27 10 AM](https://github.com/erc6900/reference-implementation/assets/8624213/f1290fe1-98b8-40a7-ae46-f7ef2732150d)|

Please let me know if you see any flaws with this!